+++ tests/variables/isset.test.mjs
@@ -1,76 +1,29 @@
-import { test, expect } from '@jest/globals';
-import isset from '../../src/variables/isset.mjs';
+import { expect, test, describe } from '@jest/globals';
+import { isset } from '../../src/variables.mjs';
 
-class stdClass {}
+describe('Функция isset', () => {
+    test('Должна возвращать true для определенных значений', () => {
+        expect(isset('example')).toBe(true);
+        expect(isset(123)).toBe(true);
+        expect(isset(true)).toBe(true);
+        expect(isset(false)).toBe(true);
+        expect(isset([1, 2, 3])).toBe(true);
+        expect(isset({ key: 'value' })).toBe(true);
+    });
 
-/**
- * https://onlinephp.io?s=pdRNa8JAEAbgeyD_YZDAJqCSpT3ViodSUSitd5GySVYNLKu4u-1B-t-bj3ayNj0E5pZ3ZnjIHHYeF-fjOQzCIPoQykmYg704OQsDmR9PwOrwAAymEJfGSBu3Ywks2h6DqrsXykiWVEOb1eb9-e1ldgM2bRSbRCVT5FIqxZHiVOpuyu9RqwMVZClDb5SOyBz3OE7nmsFOrCMdbfud2mQ664l0zNhLqQ8e2RbI8HaH5HZHxbT8BGOLJyWMQfa3QH7TTue2PGmIE7h-dY_7pkzewCmFdB2o4Pp1iV71TeUmvjchgOLnDkXZ3635GPqLizFE2SA49yGnC7kvtSz6Wj5Iy3rHvFL_uecD_-8b&v=8.3.4
- */
-test('isset', () => {
-    // php true: true
-    expect(isset(true)).toBeTruthy();
+    test('Должна возвращать false для null и undefined', () => {
+        expect(isset(null)).toBe(false);
+        expect(isset(undefined)).toBe(false);
+    });
 
-    // php false: true
-    expect(isset(false)).toBeTruthy();
+    test('Должна возвращать false, если хотя бы одно значение null или undefined', () => {
+        expect(isset('example', null)).toBe(false);
+        expect(isset(123, undefined)).toBe(false);
+        expect(isset(true, null, undefined)).toBe(false);
+    });
 
-    // php 0: true
-    expect(isset(0)).toBeTruthy();
-
-    // php 1: true
-    expect(isset(1)).toBeTruthy();
-
-    // php 3.14: true
-    expect(isset(3.14)).toBeTruthy();
-
-    // php "0": true
-    expect(isset('0')).toBeTruthy();
-
-    // php "1": true
-    expect(isset('1')).toBeTruthy();
-
-    // php "true": true
-    expect(isset('true')).toBeTruthy();
-
-    // php "false": true
-    expect(isset('false')).toBeTruthy();
-
-    // php "": true
-    expect(isset('')).toBeTruthy();
-
-    // php "string": true
-    expect(isset('string')).toBeTruthy();
-
-    // php []: true
-    expect(isset([])).toBeTruthy();
-    expect(isset({})).toBeTruthy();
-
-    // php stdClass: true
-    expect(isset(new stdClass())).toBeTruthy();
-
-    // php function () {}: true
-    expect(isset(() => {})).toBeTruthy();
-
-    // php null: false
-    expect(isset(null)).toBeFalsy();
-
-    // php INF: true
-    expect(isset(Infinity)).toBeTruthy();
-
-    // php -INF: true
-    expect(isset(-Infinity)).toBeTruthy();
-
-    // php 1, null: false
-    expect(isset(1, null)).toBeFalsy();
-
-    // php 1, undefined: false
-    expect(isset(1, undefined)).toBeFalsy();
-
-    // php 1, false
-    expect(isset(1, false)).toBeTruthy();
-
-    // js
-    expect(isset(undefined)).toBeFalsy();
-    expect(isset(NaN)).toBeTruthy();
-    expect(isset(stdClass)).toBeTruthy();
-    expect(isset(Symbol())).toBeTruthy();
+    test('Должна возвращать true, если все значения определены', () => {
+        expect(isset('example', 123, true)).toBe(true);
+        expect(isset([1, 2, 3], { key: 'value' })).toBe(true);
+    });
 });